tune: Remove all compilation warnings

+++ src/minhook/src/trampoline.c
@@ -280,7 +280,7 @@ BOOL CreateTrampolineFunction(PTRAMPOLINE ct)
 #else
         __movsb((LPBYTE)ct->pTrampoline + newPos, pCopySrc, copySize);
 #endif
-        newPos += copySize;
+        newPos += (UINT8)copySize;
         oldPos += hs.len;
     }
     while (!finished);
